Split HM integration tests to deal with the 10min Travis timout
Travis build will be terminated if there hasn't been any log output
in 10 min. Latest tests added to HM integration spec had increased
the build time, so every build is failing due to the timeout.
