+++ spec/support/integration_example_group.rb
@@ -54,6 +54,36 @@ module IntegrationExampleGroup
     format_output(run_bosh(cmd)).should == format_output(expected_output)
   end
 
+  def get_vms
+    output = run_bosh("vms --details")
+    table = output.lines.grep(/\|/)
+
+    table = table.map { |line| line.split('|').map(&:strip).reject(&:empty?) }
+    headers = table.shift || []
+    headers.map! do |header|
+      header.downcase.tr('/ ', '_').to_sym
+    end
+    output = []
+    table.each do |row|
+      output << Hash[headers.zip(row)]
+    end
+    output
+  end
+
+  def wait_for_vm(name)
+    5.times do
+      vm = get_vms.detect { |v| v[:job_index] == name }
+      return vm if vm
+    end
+    nil
+  end
+
+  def kill_job_agent(name)
+    vm = get_vms.detect { |v| v[:job_index] == name }
+    Process.kill('INT', vm[:cid].to_i)
+    vm[:cid]
+  end
+
   def self.included(base)
     base.before(:each) do |example|
       unless $sandbox_started
